+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix inability to build duckdb on Alpine Linux in some circumstances
+
 ## [1.4.0] - 2025-04-16
 
 ### Added