+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] - 2025-04-17
+
 ### Changed
 
 - Support linux/amd64 and linux/arm64 platforms for containers
@@ -86,7 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Release
 
-[Unreleased]: https://github.com/yellow-corps/ibis/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/yellow-corps/ibis/compare/v1.4.1...HEAD
+[1.4.1]: https://github.com/yellow-corps/ibis/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/yellow-corps/ibis/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/yellow-corps/ibis/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/yellow-corps/ibis/compare/v1.3.0...v1.3.1